backend/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
pydantic
passlib
bcrypt==4.0.1
python-jose[cryptography]
python-multipart